requirements.txt
youtube-transcript-api==0.6.1
tqdm==4.66.1
pandas==1.5.3
Flask==2.2.3
gunicorn==21.2.0
Werkzeug==2.2.2
isodate==0.6.1
python-dotenv==1.0.0
pydantic==2.5.2
pydantic_core==2.14.5
openapi-schema-pydantic==1.2.4
instructor==0.4.5
openai==1.1.1
